requirements.txt
Faker
kafka-python==2.0.2
requests==2.25.1